+++ src/test/java/blackjack/domain/participant/PlayersTest.java
@@ -54,7 +54,7 @@ public class PlayersTest {
     @DisplayName("플레이어의 이름, 플레이어의 베팅 금액을 동시에 입력받는 정적 팩토리 메서드 테스트")
     void initFactoryMethodWithTwoParameter() {
         List<String> playerName = Arrays.asList("joel", "bada", "jon");
-        List<Double> playerMoney = Arrays.asList(1000.0, 2000.0, 3000.0);
+        List<Integer> playerMoney = Arrays.asList(1000, 2000, 3000);
         assertThatCode(() -> Players.of(playerName, playerMoney))
                 .doesNotThrowAnyException();
     }
@@ -63,7 +63,7 @@ public class PlayersTest {
     @DisplayName("플레이어의 베팅 금액에 0 미만인 값이 포함되어 있으면 에러가 발생한다")
     void belowZeroMoneyInit() {
         List<String> playerName = Arrays.asList("joel", "bada", "jon");
-        List<Double> playerMoney = Arrays.asList(1000.0, 2000.0, -3000.0);
+        List<Integer> playerMoney = Arrays.asList(1000, 2000, -3000);
         assertThatThrownBy(() -> Players.of(playerName, playerMoney))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("베팅 금액은 0 이상이여야 합니다.");
@@ -73,7 +73,7 @@ public class PlayersTest {
     @DisplayName("플레이어의 이름, 플레이어의 베팅 금액을 담은 리스트의 크기가 다르면 에러가 발생한다")
     void listSizeDifferenceInit() {
         List<String> playerName = Arrays.asList("joel", "bada", "jon");
-        List<Double> playerMoney = Arrays.asList(1000.0, 2000.0);
+        List<Integer> playerMoney = Arrays.asList(1000, 2000);
         assertThatThrownBy(() -> Players.of(playerName, playerMoney))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("입력 받은 플레이어의 수와 베팅 금액의 수가 같지 않습니다");
